1. 앨범 블록 추가
