+++ client/src/app/board/{components}/block/extra/{components}/defaultLabel.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const DefaultLabel = () => {
+    return (
+        <div className={'absolute top-2 right-2 z-10 py-1 px-2 bg-blue-400 text-white'}>
+            <span className={'flex text-xs'}>
+                대표
+            </span>
+        </div>
+    )
+}
+
+export default DefaultLabel;
