+++ client/src/app/board/{services}/types.ts
@@ -66,11 +66,12 @@ export class Category {
     public readonly roles: RoleType[];
 
     public static readonly list = [
-        new Category("0", "전체보기", [RoleType.ADMIN, RoleType.USER, RoleType.MASTER, RoleType.GUEST]),
-        new Category("1", "공지게시판", [RoleType.ADMIN, RoleType.MASTER]),
-        new Category("2", "자유게시판", [RoleType.ADMIN, RoleType.USER, RoleType.MASTER]),
+        new Category("0", "전체 보기", [RoleType.ADMIN, RoleType.USER, RoleType.MASTER, RoleType.GUEST]),
+        new Category("1", "공지 게시판", [RoleType.ADMIN, RoleType.MASTER]),
+        new Category("2", "자유 게시판", [RoleType.ADMIN, RoleType.USER, RoleType.MASTER]),
         new Category("3", "Q & A", [RoleType.ADMIN, RoleType.USER, RoleType.MASTER]),
-        new Category("4", "알뜰게시판", [RoleType.ADMIN, RoleType.USER, RoleType.MASTER]),
+        new Category("4", "알뜰 게시판", [RoleType.ADMIN, RoleType.USER, RoleType.MASTER]),
+        new Category("5", "이미지 게시판", [RoleType.ADMIN, RoleType.USER, RoleType.MASTER]),
     ];
 
     constructor(id: string, name: string, roles: RoleType[] = []) {
