+++ server/src/main/java/com/anamensis/server/dto/request/FileRequest.java
@@ -6,6 +6,5 @@ public class FileRequest {
         long tableCodePk,
 
         long categoryPk
-
     ) {}
 }
